Fixed selection of multiple layers leaving artefacts.
- Reason:
In simulation.ts the group was set to use clustering.
This means all features of all layers in the group are combined into one array in csWeb. When then using features with overlapping id's means it will leave leaflets on the map.

Solutions: Use unique id's for all features, or don't use clustering